+++ np.Templating/CHANGELOG.md
@@ -17,6 +17,7 @@ DBW: REMEMBER THAT IF YOU ADDED ANY HELPERS IMPORTS, ADD THEM TO THE HELPER MODU
 - Add ability to skip AI error analysis by adding `disableAIErrorAnalysis: true` to the frontmatter of the template
 - Update templateRunner to allow for passing an object as args (e.g. `templateRunnerExecute('templateName', true, { getNoteTitled: 'myNoteTitle' })`)
 - Add `headingLevel` and `addHeadingLocation` to templateRunner to allow for control over where the heading is added and what level it is
+- add `tp:help` command to open templating help page in browser
 
 ## [2.0.20] 2025-08-07 @dwertheimer
 - Fix folder selection in new note command (adds folder path to chooser)